backend/requirements.txt
fastapi
uvicorn[standard]
langchain
sentence-transformers
faiss-cpu
pymupdf
sqlmodel
supabase
python-multipart
python-dotenv
